requirements.txt
setuptools
playingcards1598>=1.0.9
numpy
pandas